requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.3
django-bootstrap-modal-forms==2.2.0
django-cache-url==3.4.2
django-colorfield==0.7.2
django-datetime-widget==0.9.3
django-environ==0.9.0
environs==9.5.0
gunicorn==20.1.0
marshmallow==3.19.0
packaging==21.3
Pillow==9.3.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
